Move getJson() to AbstractLoader for base class usage
